requirements.txt
# RuView Scan - Python Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pyyaml>=6.0
click>=8.1.0
numpy>=1.24.0
scipy>=1.11.0
pydantic>=2.0.0
